oo-919 branch repair
This branch was revieved by Marko, and I made changes requested by him.
